refactor(dirs): refactor directory structure (#64)
