+++ autotest/test_base.py
@@ -4,9 +4,9 @@ import numpy as np
 import pytest
 
 from pynhm.base.accessor import Accessor
-from pynhm.base.Time import Time
 from pynhm.base.control import Control
 from pynhm.base.storageUnit import StorageUnit
+from pynhm.base.Time import Time
 from pynhm.utils.parameters import PrmsParameters
