+++ pynhm/__init__.py
@@ -1,11 +1,13 @@
+from pynhm.hydrology.PRMSCanopy import PRMSCanopy
+from pynhm.hydrology.PRMSGroundwater import PRMSGroundwater
+
 from .atmosphere.AtmBoundaryLayer import AtmBoundaryLayer
 from .atmosphere.NHMBoundaryLayer import NHMBoundaryLayer
 from .atmosphere.NHMSolarGeometry import NHMSolarGeometry
+from .base.accessor import Accessor
+from .base.control import Control
 from .base.StateAccess import StateAccess
 from .base.Time import Time
-from .canopy.PRMSCanopy import PRMSCanopy
-from .groundwater.PRMSGroundwater import PRMSGroundwater
-from .groundwater.PRMSGroundwater_better import PRMSGroundwaterBetter
 from .preprocess.cbh import CBH
 from .preprocess.csv_utils import CsvFile
 from .pynhm import driver
@@ -16,18 +18,13 @@ from .utils import (
     NetCdfWrite,
     PrmsParameters,
 )
-from .variableClass import VariableFromNetcdf
 from .version import __author__, __author_email__, __version__
 
 __all__ = [
     "atmosphere",
     "base",
-    "boundary_conditions",
-    "canopy",
-    "groundwater",
+    "hydrology",
     "preprocess",
-    "runoff",
     "utils",
     "driver",
-    "variableClass",
 ]
