--- pynhm/boundary_conditions/__init__.py
@@ -1,1 +1,0 @@
-from .boundaryConditions import BoundaryConditions
