--- pynhm/runoff/__init__.py
@@ -1,1 +1,0 @@
-from .prmsSurfaceRunoff import prmsSurfaceRunoff
